+++ process.h
@@ -22,7 +22,7 @@ typedef struct {
 } Process;
 
 Process *get_processes(DIR** procdir, struct dirent** nextprocdir, int *out_count);
-void add_process(Process **proc_arr, size_t *count, size_t *capacity, Process p);
+void add_process(Process **proc_arr, int *count, size_t *capacity, Process p);
 void tokenize_data(char *stat_str, char **fields);
 bool check_if_process(char *dir_name);
 FILE *get_stat_file(char *pid);
